refactor(plugin-ui): extract tasks description + rename config
