+++ packages/vue-cli-plugin-styleguidist/ui/config.js
