+++ packages/vue-cli-plugin-styleguidist/ui/tasks.js
@@ -1,0 +1,13 @@
+export default function describeTasks(api) {
+	api.describeTask({
+		match: /vue-cli-service styleguidist$/,
+		description: 'Compiles and hot-reloads styleguide for development',
+		link: 'https://vue-styleguidist.github.io/CLI.html#cli-commands-and-options'
+	})
+
+	api.describeTask({
+		match: /vue-cli-service styleguidist:build$/,
+		description: 'Compiles styleguide assets for production',
+		link: 'https://vue-styleguidist.github.io/CLI.html#cli-commands-and-options'
+	})
+}
